Split tests/test_holiday_groups.py (#2803)
