+++ tests/groups/test_christian.py
@@ -1,0 +1,37 @@
+#  holidays
+#  --------
+#  A fast, efficient Python library for generating country, province and state
+#  specific sets of holidays on the fly. It aims to make determining whether a
+#  specific date is a holiday as fast and flexible as possible.
+#
+#  Authors: Vacanza Team and individual contributors (see CONTRIBUTORS file)
+#           dr-prodigy <dr.prodigy.github@gmail.com> (c) 2017-2023
+#           ryanss <ryanssdev@icloud.com> (c) 2014-2017
+#  Website: https://github.com/vacanza/holidays
+#  License: MIT (see LICENSE file)
+
+from unittest import TestCase
+
+from holidays.groups import ChristianHolidays
+from holidays.holiday_base import HolidayBase
+
+
+class TestChristianHolidays(TestCase):
+    def test_check_calendar(self):
+        self.assertRaises(ValueError, lambda: ChristianHolidays("INVALID_CALENDAR"))
+
+    def test_add_christmas_day_three(self):
+        class TestHolidays(HolidayBase, ChristianHolidays):
+            def __init__(self, *args, **kwargs):
+                ChristianHolidays.__init__(self)
+                super().__init__(*args, **kwargs)
+
+            def _populate(self, year):
+                super()._populate(year)
+
+        test_holidays = TestHolidays()
+
+        test_holidays._populate(2022)
+        test_holidays._add_christmas_day_three("Third day")
+        self.assertIn("2022-12-27", test_holidays)
+        self.assertEqual(1, len(test_holidays))
